Ticket #10501
