+++ api/controllers/inputs/workspace.controllers.js
@@ -1,4 +1,7 @@
+const JSZip = require('jszip')
+const fs = require('fs')
 const { Workspace, Template, Component, User } = require('../../database.js')
+const path = require('path')
 const componentsList = require('../toCreate.js')
 
 const addWorkspace = async (req, res, next) => {
@@ -204,54 +207,49 @@ const retrieveWorkspace = async (id) => {
 }
 
 const addWorkspaceImport = async (req, res) => {
-  // const userEmail = req.params.userEmail
-  const { body } = req.body
-  console.log(body)
+  const file = req.file
+  console.log(file)
+
+  // Verifica si se cargó un archivo
+  if (!file) {
+    return res
+      .status(400)
+      .json({ error: 'No se ha proporcionado ningún archivo' })
+  }
+
   try {
-    // const userFound = await User.findOne({
-    //   where: {
-    //     email: userEmail,
-    //     isDeleted: false
-    //   }
-    // })
-    // if (!userFound) throw new Error('User not found or banned')
-    // const homepage = await Component.create(componentsList.Home)
-    // const newTemplate = await Template.create()
-    // const newWorkspace = await Workspace.create()
+    // Lee el contenido del archivo ZIP
+    const data = await fs.promises.readFile(file.path)
 
-    // await newTemplate.addPages(homepage)
-    // await newWorkspace.addProjects(newTemplate)
-    // await userFound.addWorkspaces(newWorkspace)
+    // Descomprime el archivo ZIP
+    const zip = await JSZip.loadAsync(data)
+    const outputFolder = path.join('uploads', file.originalname)
 
-    // const userUpdated = await User.findOne({
-    //   where: {
-    //     email: userEmail,
-    //     isDeleted: false
-    //   },
-    //   include: {
-    //     model: Workspace,
-    //     as: 'workspaces',
-    //     include: {
-    //       model: Template,
-    //       as: 'projects',
-    //       include: [
-    //         {
-    //           model: Component,
-    //           as: 'pages'
-    //         },
-    //         {
-    //           model: Component,
-    //           as: 'components'
-    //         }
-    //       ]
-    //     }
-    //   }
-    // })
+    // Crea la carpeta de salida si no existe
+    if (!fs.existsSync('uploads')) {
+      fs.mkdirSync('uploads')
+    }
 
-    res.json(body)
+    // Extrae cada archivo del ZIP y guárdalo en la carpeta
+    await Promise.all(
+      Object.entries(zip.files).map(async ([relativePath, file]) => {
+        if (!file.dir) {
+          const filePath = path.join(outputFolder, relativePath)
+          const content = await file.async('nodebuffer')
+          fs.writeFileSync(filePath, content)
+          console.log(`Archivo descomprimido: ${filePath}`)
+        }
+      })
+    )
+
+    return res
+      .status(200)
+      .json({ message: 'Archivo ZIP descomprimido exitosamente' })
   } catch (error) {
-    console.log(error)
-    return res.status(200).send(error)
+    console.error('Error al descomprimir el archivo ZIP:', error)
+    return res
+      .status(500)
+      .json({ error: 'Error al descomprimir el archivo ZIP' })
   }
 }
