+++ src/Pages/dashboard/Editor/mainheader/Import.js
@@ -16,7 +16,7 @@ export default function Import({ toggleModalImport }) {
         formData.append('file', file)
 
         axios
-          .post('/import', formData, {
+          .post('/workspace/import', formData, {
             headers: {
               'Content-Type': 'multipart/form-data'
             }
